+++ packages/miniflare/src/config/v4-schema.ts
@@ -197,6 +197,12 @@ const V4UnsafeDirectSocketSchema = z.object({
 	proxy: z.boolean().optional(),
 });
 
+const V4ConnectHandlerSchema = z.object({
+	protocol: z.enum(["tcp"]),
+	port: z.number(),
+	address: z.string().optional(),
+});
+
 const V4IdEntrySchema = z.object({
 	id: z.string(),
 	remoteProxyConnectionString: RemoteProxyConnectionStringSchema.optional(),
@@ -364,11 +370,11 @@ const V4WorkerOptionsShapeSchema = z.object({
 	outboundService: V4ServiceDesignatorSchema.optional(),
 	unsafeEphemeralDurableObjects: z.boolean().optional(),
 	unsafeDirectSockets: z.array(V4UnsafeDirectSocketSchema).optional(),
+	connectHandlers: z.array(V4ConnectHandlerSchema).optional(),
 	unsafeOverrideFetchWorker: z.string().optional(),
 	unsafeEvalBinding: z.string().optional(),
 	unsafeUseModuleFallbackService: z.boolean().optional(),
 	unsafeRegisterWorker: z.boolean().optional(),
-	hasAssetsAndIsVitest: z.boolean().optional(),
 	tails: z.array(V4ServiceDesignatorSchema).optional(),
 	streamingTails: z.array(V4ServiceDesignatorSchema).optional(),
 	stripCfConnectingIp: z.boolean().default(true),
@@ -780,11 +786,15 @@ export type V4WorkerOptionsShape = {
 		entrypoint?: string;
 		proxy?: boolean;
 	}>;
+	connectHandlers?: Array<{
+		protocol: "tcp";
+		port: number;
+		address?: string;
+	}>;
 	unsafeOverrideFetchWorker?: string;
 	unsafeEvalBinding?: string;
 	unsafeUseModuleFallbackService?: boolean;
 	unsafeRegisterWorker?: boolean;
-	hasAssetsAndIsVitest?: boolean;
 	tails?: V4ServiceDesignator[];
 	streamingTails?: V4ServiceDesignator[];
 	stripCfConnectingIp?: boolean;